Added periodic authorization check Clean up
