+++ views/server-config/action-form-layout.js
@@ -1,10 +1,8 @@
 import React, {useCallback, useState} from 'react'
 import {UtcTimestamp} from '@stellar-expert/ui-framework'
 
-export function updateTimeValidation(settings) {
-    const timestamp = settings.timestamp
-    const expirationDate = settings.expirationDate
-    const minDate = settings.config.minDate
+export function updateTimeValidation({timestamp, expirationDate, ...config}) {
+    const minDate = config.minDate
     const min = new Date().getTime() + 30 * 60 * 1000
     const max = new Date().getTime() + 10 * 24 * 60 * 60 * 1000
     if (timestamp === 0 && minDate === 0 && expirationDate)
